Changed addByteToFile() to take int parameter.
Since every call to addByteToFile() is called with a hex integer
argument, it makes more sense to do this. Primitive types can't be
implicitly cast up to an object like Object so it was throwing lots of
exceptions.

Also removed else clause from 'if (isFunctionKey())' in begin(). This
was throwing an error for each command since it fails for everything but
a function key.

+++ Encoder/src/org/hak5/ducky/command/String.java
@@ -57,7 +57,7 @@ public class String implements Command {
                 }
             }
 
-            state.addByteToFile(shiftByte);
+            state.addByteToFile((int) shiftByte);
         }
     }
